+++ src/framework/piral-core/src/utils/index.ts
@@ -1,3 +1,5 @@
+export { isfunc } from 'piral-base';
+
 export * from './compare';
 export * from './data';
 export * from './extension';