+++ src/plugins/piral-axios/src/types.ts
@@ -1,5 +1,5 @@
 import type {} from 'piral-core';
-import { AxiosInstance } from 'axios';
+import type { AxiosInstance } from 'axios';
 
 declare module 'piral-core/lib/types/custom' {
   interface PiletCustomApi extends PiletAxiosApi {}