+++ src/plugins/piral-dashboard/src/components.tsx
@@ -1,7 +1,6 @@
-import * as React from 'react';
+import { ComponentType } from 'react';
 import { getPiralComponent } from 'piral-core';
 import { DashboardContainerProps, DashboardTileProps } from './types';
 
-export const PiralDashboardContainer: React.ComponentType<DashboardContainerProps> =
-  getPiralComponent('DashboardContainer');
-export const PiralDashboardTile: React.ComponentType<DashboardTileProps> = getPiralComponent('DashboardTile');
+export const PiralDashboardContainer: ComponentType<DashboardContainerProps> = getPiralComponent('DashboardContainer');
+export const PiralDashboardTile: ComponentType<DashboardTileProps> = getPiralComponent('DashboardTile');