+++ src/plugins/piral-menu/src/components.tsx
@@ -1,6 +1,6 @@
-import * as React from 'react';
+import type { ComponentType } from 'react';
 import { getPiralComponent } from 'piral-core';
 import { MenuContainerProps, MenuItemProps } from './types';
 
-export const PiralMenuContainer: React.ComponentType<MenuContainerProps> = getPiralComponent('MenuContainer');
-export const PiralMenuItem: React.ComponentType<MenuItemProps> = getPiralComponent('MenuItem');
+export const PiralMenuContainer: ComponentType<MenuContainerProps> = getPiralComponent('MenuContainer');
+export const PiralMenuItem: ComponentType<MenuItemProps> = getPiralComponent('MenuItem');