+++ src/plugins/piral-modals/src/components.tsx
@@ -1,6 +1,6 @@
-import * as React from 'react';
+import { ComponentType } from 'react';
 import { getPiralComponent } from 'piral-core';
 import { ModalsHostProps, ModalsDialogProps } from './types';
 
-export const PiralModalsHost: React.ComponentType<ModalsHostProps> = getPiralComponent('ModalsHost');
-export const PiralModalsDialog: React.ComponentType<ModalsDialogProps> = getPiralComponent('ModalsDialog');
+export const PiralModalsHost: ComponentType<ModalsHostProps> = getPiralComponent('ModalsHost');
+export const PiralModalsDialog: ComponentType<ModalsDialogProps> = getPiralComponent('ModalsDialog');