+++ src/plugins/piral-notifications/src/types.ts
@@ -1,5 +1,5 @@
-import { ReactElement, ComponentType } from 'react';
-import { Disposable, BaseComponentProps, AnyComponent } from 'piral-core';
+import type { ReactElement, ComponentType } from 'react';
+import type { Disposable, BaseComponentProps, AnyComponent } from 'piral-core';
 
 declare module 'piral-core/lib/types/custom' {
   interface PiletCustomApi extends PiletNotificationsApi {}