+++ src/plugins/piral-redux/src/types.ts
@@ -1,6 +1,6 @@
 import type {} from 'piral-core';
-import { Store, Action, Reducer, Dispatch, AnyAction } from 'redux';
-import { ComponentType, FC } from 'react';
+import type { ComponentType, FC } from 'react';
+import type { Store, Action, Reducer, Dispatch, AnyAction } from 'redux';
 
 declare module 'piral-core/lib/types/custom' {
   interface PiletCustomApi extends PiletReduxApi {}