+++ src/plugins/piral-search/src/components.tsx
@@ -1,7 +1,7 @@
-import * as React from 'react';
+import type { ComponentType } from 'react';
 import { getPiralComponent } from 'piral-core';
 import { SearchContainerProps, SearchInputProps, SearchResultProps } from './types';
 
-export const PiralSearchContainer: React.ComponentType<SearchContainerProps> = getPiralComponent('SearchContainer');
-export const PiralSearchInput: React.ComponentType<SearchInputProps> = getPiralComponent('SearchInput');
-export const PiralSearchResult: React.ComponentType<SearchResultProps> = getPiralComponent('SearchResult');
+export const PiralSearchContainer: ComponentType<SearchContainerProps> = getPiralComponent('SearchContainer');
+export const PiralSearchInput: ComponentType<SearchInputProps> = getPiralComponent('SearchInput');
+export const PiralSearchResult: ComponentType<SearchResultProps> = getPiralComponent('SearchResult');