+++ src/plugins/piral-translate/src/default.tsx
@@ -1,4 +1,5 @@
+import type { FC } from 'react';
 import { defaultRender } from 'piral-core';
 import { LanguagesPickerProps } from './types';
 
-export const DefaultPicker: React.FC<LanguagesPickerProps> = (props) => defaultRender(undefined);
+export const DefaultPicker: FC<LanguagesPickerProps> = (props) => defaultRender(undefined);