+++ src/plugins/piral-update/src/components.tsx
@@ -1,5 +1,5 @@
-import * as React from 'react';
+import type { ComponentType } from 'react';
 import { getPiralComponent } from 'piral-core';
 import { UpdateDialogProps } from './types';
 
-export const PiralUpdateDialog: React.ComponentType<UpdateDialogProps> = getPiralComponent('UpdateDialog');
+export const PiralUpdateDialog: ComponentType<UpdateDialogProps> = getPiralComponent('UpdateDialog');