+++ src/plugins/piral-update/src/default.tsx
@@ -1,5 +1,5 @@
-import * as React from 'react';
+import { FC } from 'react';
 import { defaultRender } from 'piral-core';
 import { UpdateDialogProps } from './types';
 
-export const DefaultUpdateDialog: React.FC<UpdateDialogProps> = (props) => defaultRender(props.children);
+export const DefaultUpdateDialog: FC<UpdateDialogProps> = (props) => defaultRender(props.children);