+++ src/tooling/piral-cli/src/apps/debug-pilet.ts
@@ -272,18 +272,22 @@ export async function debugPilet(baseDir = process.cwd(), options: DebugPiletOpt
   }
 
   const publicUrl = '/';
+  const { pilet: piletInjector, ...otherInjectors } = krasConfig.injectors;
   const injectorConfig = {
+    meta: 'debug-meta.json',
+    feed,
+    ...piletInjector,
     active: true,
     pilets,
     app: appDir,
     handle: ['/', api],
     api,
-    feed,
   };
 
   krasConfig.map['/'] = '';
   krasConfig.map[api] = '';
-  krasConfig.injectors = reorderInjectors(injectorName, injectorConfig, krasConfig.injectors);
+
+  krasConfig.injectors = reorderInjectors(injectorName, injectorConfig, otherInjectors);
 
   log('generalVerbose_0004', `Using kras with configuration: ${JSON.stringify(krasConfig, undefined, 2)}`);