+++ src/tooling/piral-cli-webpack/src/index.ts
@@ -1,8 +1,0 @@
-import * as actions from './actions';
-import { CliPlugin } from 'piral-cli';
-
-const plugin: CliPlugin = (cli) => {
-  cli.withBundler('webpack', actions);
-};
-
-module.exports = plugin;