+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Removed support for IE11 (#467)
 - Added new convenience package `piral-hooks-utils`
 - Added new npm script running bundler `piral-cli-xbuild` (#470)
+- Added types to `piral-cli-webpack/extend-config` and `piral-cli-webpack5/extend-config`
 
 ## 0.14.9 (January 21, 2022)