+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Added new convenience package `piral-hooks-utils`
 - Added new npm script running bundler `piral-cli-xbuild` (#470)
 - Added types to `piral-cli-webpack/extend-config` and `piral-cli-webpack5/extend-config`
+- Added semver possibilities for resolving shared dependencies
 
 ## 0.14.9 (January 21, 2022)