+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.15.0 (tbd)
 
 - Improved support and documentation for wrappers
+- Improved default metadata using `piletConfig` from *package.json* in `pilet debug` (#458)
 - Moved templates to dedicated repository (#458)
 - Changed the default bundler to be `piral-cli-webpack5` (#469)
 - Change webpack configs to use `oneOf` for assets (#451)
@@ -10,6 +11,7 @@
 - Removed support for IE11 (#467)
 - Added new prop `meta` to registered pages
 - Added pre-bundled minified versions to each plugin package
+- Added support for *debug-meta.json* when using `pilet debug` (#462)
 - Added new convenience package `piral-hooks-utils`
 - Added new npm script running bundler `piral-cli-xbuild` (#470)
 - Added types to `piral-cli-webpack/extend-config` and `piral-cli-webpack5/extend-config`